Implement Functions.matchN
Sometimes it is helpful to be able to call an Op with an array of
arguments, however we were before only able to do so when we had them
individually. We now have matchN, which is able to return a FunctionN
(which wraps some other arity Function) that can work on an array of
arguments.